services/image_service.py
import os
import base64
from PIL import Image
from io import BytesIO
import uuid

class ImageService:
    """Service for retrieving and processing images."""

    def retrieve_images(self, data: str) -> list:
        # If LLM returns references to images, process them here
        # For now, just return an empty list or process base64 images if given
        return []

def save_base64_image(encoded_image: str) -> str:
    image_data = base64.b64decode(encoded_image)
    image = Image.open(BytesIO(image_data))
    image_path = f"/tmp/image_{uuid.uuid4()}.png"
    image.save(image_path)
    return image_path
